+++ seleniumide/src/main/java/org/musetest/seleniumide/steps/SendKeysConverter.java
@@ -20,7 +20,7 @@ public class SendKeysConverter implements StepConverter
             return null;
         StepConfiguration step = new StepConfiguration(SendKeys.TYPE_ID);
         step.addSource(SendKeys.ELEMENT_PARAM, LocatorConverters.get().convert(param1));
-        step.addSource(SendKeys.KEYS_PARAM, ValueConverters.get().convertValue(param2));
+        step.addSource(SendKeys.KEYS_PARAM, ValueConverters.get().convert(param2));
         if (TYPE.equals(command))
             step.addSource(SendKeys.CLEAR_PARAM, ValueSourceConfiguration.forValue(true));
         return step;