+++ seleniumide/src/test/java/org/musetest/seleniumide/tests/mocks/MockTestConverter.java
@@ -1,0 +1,18 @@
+package org.musetest.seleniumide.tests.mocks;
+
+import org.musetest.seleniumide.*;
+
+import java.io.*;
+
+/**
+ * @author Christopher L Merrill (see LICENSE.txt for license details)
+ */
+public class MockTestConverter extends TestConverter
+    {
+    public MockTestConverter()
+        {
+        super(new ByteArrayInputStream("unused".getBytes()));
+        }
+    }
+
+